GROUNDWORK-1688 Split libtransit for agent
